+++ goap/Take.cpp
@@ -146,28 +146,28 @@ Take::operator std::string()
 	return "Take";
 }
 
-int Take::Cost(RoomManager* rm)
+float Take::Cost(RoomManager* rm)
 {
 	// return a measure of
 	// 1 - if that thing doesn't BELONG to you
 	// 2 - bonus points if it BELONGS to someone else, and not a room
 
-	int cost;
+	float cost;
 
 	auto _agent = GetArg(SemanticRole::AGENT);
 	auto _patient = GetArg(SemanticRole::PATIENT);
 	
 	if(_patient->instance->GetOwner() == 0)
 	{
-		cost = 10;
+		cost = 10.0f;
 	}
 	else if(_patient->instance->GetOwner() != _agent->instance)
 	{
-		cost = 50;
+		cost = 50.0f;
 	}
 	else if(_patient->instance->GetOwner() == _agent->instance)
 	{
-		cost = 0;
+		cost = 0.0f;
 	}
